Modularize navigation events tracking
