+++ src/extension.js
@@ -8,6 +8,8 @@ const path = require('path');
 const Terminal = require('./terminal');
 const activeWindow = require('active-win');
 const { checkBashProfilePath, checkPowerShellProfilePath } = require('./profileHelpers');
+const { debounce, getCurrentDir, user, hostname } = require('./helpers');
+const navigation = require('./navigation');
 
 var tracker = null;
 var iter = 0;
@@ -31,18 +33,7 @@ var hostname = os.hostname();
 var terminalList;
 var terminalInstance;
 var previousDocument = null;
-var navigationHistory = [];
 var selectionHistory = [];
-const debounceTimers = new Map(); // For debouncing per document
-
-function getCurrentDir() {
-	if (vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders.length > 0) {
-		return vscode.workspace.workspaceFolders[0].uri.fsPath;
-	} else {
-		vscode.window.showErrorMessage("Working folder not found, open a folder and try again.");
-		return null; // No directory found
-	}
-}
 
 /**
  * @param {vscode.ExtensionContext} context
@@ -57,12 +48,12 @@ function activate(context) {
         // Generate a unique key for the editor, e.g., using its document URI
         const key = event.textEditor.document.uri.fsPath;
         // Apply debouncing per editor
-        const debouncedHandleVisibleRangeChange = debounce(handleVisibleRangeChange, 500, key);
+        const debouncedHandleVisibleRangeChange = debounce(navigation.handleVisibleRangeChange, 500, key);
         debouncedHandleVisibleRangeChange(event.textEditor); // Pass the editor instance directly
     }));
 
     // Listen for document changes to track navigation between documents
-    context.subscriptions.push(vscode.window.onDidChangeActiveTextEditor(handleActiveTextEditorChange));
+    context.subscriptions.push(vscode.window.onDidChangeActiveTextEditor(navigation.handleActiveTextEditorChange));
 
 	// Listen for selection changes with debounced handling
 	context.subscriptions.push(vscode.window.onDidChangeTextEditorSelection(event => {
@@ -1015,132 +1006,6 @@ function containsIPAddresses(url) {
 	return ipv4Pattern.test(url) || ipv6Pattern.test(url);
 }
 
-function removeBackspaces(str) {
-	var patternString = (
-        "[\\u0000]|[\\u0001]|[\\u0002]|[\\u0003]|[\\u0004]|" +
-        "[\\u0005]|[\\u0006]|[\\u0007]|[\\u0008]|[\\u000b]|" +
-        "[\\u000c]|[\\u000d]|[\\u000e]|[\\u000f]|[\\u0010]|" +
-        "[\\u0011]|[\\u0012]|[\\u0013]|[\\u0014]|[\\u0015]|" +
-        "[\\u0016]|[\\u0017]|[\\u0018]|[\\u0019]|[\\u001a]|" +
-        "[\\u001b]|[\\u001c]|[\\u001d]|[\\u001e]|[\\u001f]|" +
-        "[\\u007f]|[\\u0040]"
-    );
-    var pattern = new RegExp(patternString, "gm");
-    while (str.indexOf("\b") !== -1) {
-        str = str.replace(/.?\x08/g, ""); // 0x08 is the ASCII code for \b
-    }
-    str = str.replace(pattern, "");  
-    return str;
-}
-
-// Debounce function improved to handle debouncing per unique key (document URI)
-function debounce(func, wait, key) {
-    return function(...args) {
-        if (!debounceTimers.has(key)) {
-            debounceTimers.set(key, null);
-        }
-        clearTimeout(debounceTimers.get(key));
-        debounceTimers.set(key, setTimeout(() => {
-            func.apply(this, args);
-        }, wait));
-    };
-}
-
-function handleVisibleRangeChange(editor) {
-    if (!editor) return;
-
-    const document = editor.document;
-    const visibleRangesInfo = editor.visibleRanges.map(range => {
-        // Calculate surrounding lines' indices, ensuring they are within document bounds
-        const startLineIndex = Math.max(range.start.line - 2, 0); // 2 lines above the first visible, or document start
-        const endLineIndex = Math.min(range.end.line + 2, document.lineCount - 1); // 2 lines below the last visible, or document end
-
-        // Extract text for the surrounding lines
-        const linesText = [];
-        
-		// Capture lines before the start of the visible range and the visible range itself
-		for (let i = startLineIndex; i <= range.end.line; i++) {
-			linesText.push(document.lineAt(i).text);
-		}
-
-        // Insert ellipsis if there's a gap indicating omitted content
-        if (range.start.line - startLineIndex > 0 || endLineIndex - range.end.line > 0) {
-            linesText.push("...");
-        }
-
-        // Capture lines at the end of the visible range and after it
-		for (let i = range.end.line; i <= endLineIndex; i++) {
-			linesText.push(document.lineAt(i).text);
-		}
-
-        let documentPath = document.uri.fsPath;
-
-        // trim the user and hostname from the document
-        if (user && hostname) {
-            let userRegex = new RegExp("\\b" + user + "\\b", "g");
-            let hostnameRegex = new RegExp("\\b" + hostname + "\\b", "g");
-            documentPath = documentPath.replace(userRegex, "user").replace(hostnameRegex, "hostname");
-        }
-
-        const entry = {
-            surroundingLines: linesText,
-            range: [range.start.line + 1, range.end.line + 1],
-            document: documentPath,
-            time: Math.floor(Date.now() / 1000),
-        };
-
-        appendNavigationEntryToFile(entry, editor);
-
-        return entry;
-    });
-
-	// Store each visible range change in navigation history
-	// Uncomment the following lines to view the visible ranges info in the console
-    // navigationHistory.push(...visibleRangesInfo);
-    // console.log('Navigation History Updated:', navigationHistory);
-}
-
-function updateVisibleEditors() {
-    vscode.window.visibleTextEditors.forEach(editor => {
-        // Directly call handleVisibleRangeChange for each visible editor
-        // Note: No need to debounce here as this is called from a debounced context or controlled events
-        handleVisibleRangeChange(editor);
-    });
-}
-
-
-function handleActiveTextEditorChange(editor) {
-	// This checks for actual editor changes, including document switches and split view adjustments
-    if (editor) {
-        let currentDocument = editor.document.uri.fsPath;
-        if (previousDocument && currentDocument !== previousDocument) {
-			// trim the user and hostname from the document
-			if (user && hostname) {
-				let userRegex = new RegExp("\\b" + user + "\\b", "g");
-				let hostnameRegex = new RegExp("\\b" + hostname + "\\b", "g");
-				currentDocument = currentDocument.replace(userRegex, "user").replace(hostnameRegex, "hostname");
-				previousDocument = previousDocument.replace(userRegex, "user").replace(hostnameRegex, "hostname");
-			}
-
-			// Store the navigation history entry for the document switch
-			let entry = {
-				document: currentDocument,
-				prevDocument: previousDocument,
-				time: Math.floor(Date.now() / 1000),
-				transition: true, // Indicates this record is about transitioning between documents
-			};
-
-			// appendNavigationEntryToFile(entry, editor);
-
-			// Uncomment the following lines to view the navigation history in the console
-			// navigationHistory.push(entry);
-			// console.log('Navigation History Updated:', navigationHistory);
-        }
-        previousDocument = currentDocument; // Update for future comparisons
-    }
-	updateVisibleEditors();
-}
-
 function handleTextEditorSelectionChange(editor) {
     // Ensure there is a valid selection
     if (!editor || !editor.selection || editor.selection.isEmpty) {
@@ -1213,26 +1078,6 @@ function handleTextEditorSelectionChange(editor) {
 	// Uncomment the following lines to view the selection history in the console 
 	// selectionHistory.push(entry);
 	// console.log('Selection History Updated:', selectionHistory);
-}
-
-function appendNavigationEntryToFile(entry, editor) {
-	const currentDir = getCurrentDir();
-    const filePath = path.join(currentDir, 'CH_cfg_and_logs', 'CH_navigation_history.ndjson');
-	const currentDocumentPath = editor.document.uri.fsPath;
-
-	// Skip appending if the current document is the navigation history file
-    if (currentDocumentPath === filePath) {
-        return;
-    }
-
-    // Append the entry to the NDJSON file
-	const ndjsonString = JSON.stringify(entry) + '\n';
-    fs.appendFile(filePath, ndjsonString, (err) => {
-        if (err) {
-            console.error('Error appending navigation entry to file:', err);
-            vscode.window.showErrorMessage('Failed to append navigation entry to file.');
-        }
-    });
 }
 
 function handleTextDocumentSave(document) {
